% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.rna.seq.R
\name{gen.rna.seq}
\alias{gen.rna.seq}
\title{Generate RNA-seq data}
\usage{
gen.rna.seq(len, meanlog = log(stats::runif(1, min = 0, max = 3)),
  sdlog = log(1 + stats::runif(1, min = 0.08, max = 1.5)))
}
\arguments{
\item{len}{number of observations to be generated}

\item{meanlog}{ln(mean)}

\item{sdlog}{ln(sd)}
}
\value{
random observations
}
\description{
From a log-normal distribution
}
\examples{
gen.rna.seq(1)
gen.rna.seq(100)
gen.rna.seq(10, log(1.2), log(1 + 0.5))
}
